converting to SFC's
