+++ src/router.ts
@@ -11,7 +11,7 @@ const router = new Router({
     {
       path: '/',
       name: 'home',
-      component: () => import('./views/home/Home.vue'),
+      component: () => import('./views/Home.vue'),
       meta: {
         requiresAuth: true,
       },
@@ -19,12 +19,12 @@ const router = new Router({
     {
       path: '/about',
       name: 'about',
-      component: () => import('./views/about/About.vue'),
+      component: () => import('./views/About.vue'),
     },
     {
       path: '/login',
       name: 'login',
-      component: () => import('./views/login/Login.vue'),
+      component: () => import('./views/Login.vue'),
     },
   ],
 });
