+++ src/views/Home.vue
@@ -1,3 +1,12 @@
+<template>
+  <div class="home">
+    <img alt="Vue logo" src="../assets/logo.png">
+    <HelloWorld msg="Welcome to Your Vue.js + TypeScript App"></HelloWorld>
+    <span>{{ message }}</span>
+  </div>
+</template>
+
+<script lang="ts">
 import { Component, Vue } from 'vue-property-decorator';
 import axios from 'axios';
 import HelloWorld from '@/components/HelloWorld.vue';
@@ -21,3 +30,4 @@ export default class Home extends Vue {
         });
     }
 }
+</script>
