+++ src/views/Login.vue
@@ -12,7 +12,35 @@
     </div>
 </template>
 
-<script src="./Login.ts"></script>
+<script lang="ts">
+import { Component, Vue } from 'vue-property-decorator';
+import authentication from '../store/AuthenticationModule';
+
+@Component
+export default class Login extends Vue {
+
+    username: string = '';
+    password: string = '';
+
+    get loginLoading() {
+        return authentication.isLoading;
+    }
+
+    get loginFailed() {
+        return authentication.isError;
+    }
+
+    submitLogin() {
+        const username = this.username;
+        const password = this.password;
+        authentication.doLogin({ username, password })
+            .then(
+                () => this.$router.push('/'),
+                () => null, // ignored
+            );
+    }
+}
+</script>
 
 <style scoped>
     .login-page {
