--- src/views/home/Home.vue
@@ -1,9 +1,0 @@
-<template>
-  <div class="home">
-    <img alt="Vue logo" src="../../assets/logo.png">
-    <HelloWorld msg="Welcome to Your Vue.js + TypeScript App"></HelloWorld>
-    <span>{{ message }}</span>
-  </div>
-</template>
-
-<script src="./Home.ts"></script>
